Update patch set 1
Patch Set 1:

It miss badly to be able to configure S-Pen offset values. Because of this I created a kernel patch and this patch to GalaxyNoteSettings.apk to get a config tab to adjust tilt offsets.

It allows to switch between right-handed/left-handed mode and fine-grained adjustment of coordinate offsets separately for all 4 possible screen orientations.

One problem, I didn't solve so far, is, that the kernel driver for the pen is not aware of the orientation of the active application. Thus, always the values for portrait orientation are used.

Maybe someone can help me along with this problem. What would be necessary is some kind of service which writes the value for the current orientation (with consideration of auto-rotation settings and application rotation behavior) to /sys/class/sec/sec_epen/epen_rotation whenever orientation changes.

Patch-set: 1